% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QeITH.r
\name{QeITH}
\alias{QeITH}
\title{quantifying intratumor immune heterogeneity}
\usage{
QeITH(data, type = c("bulk", "single-cell", "spatial"))
}
\arguments{
\item{data}{A data.frame of mRNA expression profile, cell types annotation data, or the proportions of cell types across samples.}

\item{type}{A character vector indicating the type of "bulk", "single-cell",'spatial'.}
}
\value{
A data frame containing sample IDs and intratumor heterogeneity scores -- ITH score
}
\description{
quantifying intratumor immune heterogeneity
}
\examples{
# Example usage:
# result_bulk <- QeITH(data, type = bulk")
# result_sc <- QeITH(data, type = "single-cell")
# result_sc <- QeITH(data, type = "spatial")
}
